% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_helper_functions.R
\name{add_labels}
\alias{add_labels}
\title{Add readable labels to columns}
\usage{
add_labels(df, dict)
}
\arguments{
\item{df}{Data frame. The data to label.}

\item{dict}{Data frame. Data dictionary with \code{field_name} and \code{field_label}.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  Data frame with labelled variables.
}\if{html}{\out{</div>}}
}
\description{
Set cleaned labels on all variables in a data frame, using a data dictionary
Define function for piping and combine label setting and cleaning
}
