% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{calc_time}
\alias{calc_time}
\title{Calculate time difference between two dates}
\usage{
calc_time(date1, date2, unit = "months")
}
\arguments{
\item{date1}{Date. First date (usually later date).}

\item{date2}{Date. Second date (earlier date).}

\item{unit}{String. One of "days", "months", or "years". (Default "months").}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   Numeric vector with the same length as date1/date2. Difference in months.
}\if{html}{\out{</div>}}
}
\description{
Calculate difference in months between two dates

In case for unit = "years", it uses the average year value of 365.25,
for unit = "months" it uses the average year value of 365.25 divided by 12
}
