% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_helper_functions.R
\name{capitalize_words}
\alias{capitalize_words}
\title{Capitalize first letter; string separated with comma}
\usage{
capitalize_words(x)
}
\arguments{
\item{x}{Character vector with words separated by commas.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Character vector with capitalized words.
}\if{html}{\out{</div>}}
}
\description{
Capitalize first letter of each word in a comma-separated string
easier reading
}
