% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_helper_functions.R
\name{clean_html_css}
\alias{clean_html_css}
\title{(Helper) Remove html and css tags}
\usage{
clean_html_css(label)
}
\arguments{
\item{label}{Character string. Input label with possible HTML tags.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   Cleaned character string.
}\if{html}{\out{</div>}}
}
\description{
Remove HTML and CSS tags from a string from the label
}
