% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_helper_functions.R
\name{clean_names}
\alias{clean_names}
\title{Clean Data dictionary variable names}
\usage{
clean_names(df)
}
\arguments{
\item{df}{Data frame whose column names should be cleaned.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   Data frame with cleaned names.
}\if{html}{\out{</div>}}
}
\description{
Clean data frame column names (to lowercase, underscores, remove special chars)
}
