% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_helper_functions.R
\name{convert_to_numeric}
\alias{convert_to_numeric}
\title{Convert to numeric with dot}
\usage{
convert_to_numeric(x)
}
\arguments{
\item{x}{Character vector or factor to convert (e.g. "1,23").}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Numeric vector.
}\if{html}{\out{</div>}}
}
\description{
Convert numbers with comma as decimal to numeric
}
