% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets_project_starter.R
\name{create_new_project}
\alias{create_new_project}
\title{Core function for creating a new project}
\usage{
create_new_project(
  project_name,
  base_path = getwd(),
  include_renv = FALSE,
  use_targets = TRUE,
  include_gitignore = TRUE,
  create_readme = TRUE,
  user = Sys.info()["user"]
)
}
\arguments{
\item{project_name}{String; Name of the project. Is used as file name.}

\item{base_path}{Base path where the project will be.
Has to be defined with \verb{set_wd(__your_path__/projects/)} before.}

\item{include_renv}{Logical; Whether a renv environment should be initialized. Default: \code{FALSE}.}

\item{use_targets}{}

\item{include_gitignore}{Logical; Whether a gitignore file should be initialized. Default: \code{TRUE}.}

\item{create_readme}{Logical; Whether a README file should be initialized. Default: \code{TRUE}.}

\item{user}{Name of the user. Will be used in README, config.yaml & cleaning_data_pipeline.R}
}
\value{
Path to the created zum erstellten Projektverzeichnis
}
\description{
This function uses helper functions to create an complete project setup.
It can be used as a function via a R-script. It uses renv & targets to
set up a reproducible workflow.
}
\examples{
set_wd("C:/Users/path/to/Projects")
create_new_project("Projekt1", include_renv = FALSE)
}
