% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{create_notebook_wd}
\alias{create_notebook_wd}
\title{Create new notebook}
\usage{
create_notebook_wd(topic = "notes")
}
\arguments{
\item{topic}{String; name of the new note file to be inserted.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{       Path of the new created notebook file.
}\if{html}{\out{</div>}}
}
\description{
Creates a new note file with the correct path in \code{Projects/Projectname/notebook} and date automatically.
}
\examples{
          `create_notebook_wd("notes")`

}
