% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_helper_functions.R
\name{create_raw_date_column}
\alias{create_raw_date_column}
\title{Create Raw Date Columns with Missing Day/Month Filled as "nk"}
\usage{
create_raw_date_column(df, date_column, month_column, year_column, condition)
}
\arguments{
\item{df}{Data frame. The DataFrame in which the date column will be created.}

\item{date_column}{Character. The name of the date column to be handled (e.g., \verb{“start_date”}).}

\item{month_column}{Character. The name of the month column (e.g., \verb{“start_date_month”}).}

\item{year_column}{Character. The name of the year column (e.g., \verb{“start_date_year”}).}

\item{condition}{A logical condition (or vector) to determine when to create or update the raw date column.
This condition can be any logical expression, such as \code{df$surv_stat == "alive"/1}.}

\item{flag_column}{Character. The name of the flag column that indicates whether a date should be set (e.g., \verb{“treatment_y_n”}).}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{             The input data frame with the added raw date column containing missing days/months as `“nk”`.
                      The new column will be named: `<date_column>_raw`.
}\if{html}{\out{</div>}}
}
\description{
This function generates a raw date column based on the provided conditions,
by combining month and year columns into a single date. If the date column
is missing, it inserts a placeholder "nk" with the respective month and year.
}
\examples{
# Example: Create raw status date column when surv_stat is "alive"
df2 <- df \%>\%
create_raw_date_column(
  date_column  = "status_date",
  month_column = "last_contact_date_month",
  year_column  = "last_contact_date_year",
  condition    = df$surv_stat == "alive")

# Example: Create raw therapy start date column when treat_y_n is 1
df <- create_raw_date_column(
  df2 = df,
  date_column = "therapy_start_date",
  month_column = "therapy_start_date_month",
  year_column = "therapy_start_date_year",
  condition = df$treat_y_n == 1
}
