% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_helper_functions.R
\name{create_results_wd}
\alias{create_results_wd}
\title{Creates WD for generated results}
\usage{
create_results_wd(analysis_type)
}
\arguments{
\item{analysis_type}{String; defines the type of the Output.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{            Directory of `Projects/Projectname/results` to safe the results here.
}\if{html}{\out{</div>}}

Example use: \code{create_results_wd("cleaning")}.
\code{openxlsx::write.xlsx(data, file = file.path(results_dir, "result_1_data.csv"))}.
}
\description{
To save the results with the current date and name in the correct directory \code{Projects/Projectname/results}.
}
