% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{create_surv_plot}
\alias{create_surv_plot}
\title{High‑level Kaplan–Meier plot wrapper}
\usage{
create_surv_plot(
  data = NULL,
  fit,
  time_unit = "Years",
  xscale = "m_y",
  scale_break = 24,
  x_end = NULL,
  title = "Kaplan-Meier curve",
  title_size = 11,
  show_tbls = TRUE,
  tbl1_pos = c(0.8, 0),
  tbl2_pos = c(0, 0),
  followup_times = NULL,
  risk_table_size = 2.7,
  palette = "black",
  ...
)
}
\arguments{
\item{data}{Optional \code{data.frame} containing all variables. used for the risk table. If omitted the
risk table is suppressed (because \code{ggsurvplot()} requires \code{data} for that panel).}

\item{fit}{\code{survfit} object. Stratification present in \code{fit} will be respected automatically.}

\item{time_unit}{String. Original time unit of the model (informational, used for axis label).}

\item{xscale}{See \code{get_xscale()}. Default = "m_y" (= from months to years).
Allowed pairs (\code{from_to}): "d_m", "d_y", "m_d", "m_y", "y_d", "y_m", "d_d", "m_m", "y_y",
or "" / NULL (= no conversion).}

\item{scale_break}{Numeric. Break points of the x-scale. Default = 24 months. In \emph{original} time unit.}

\item{x_end}{Numeric. Optional. Gets calculated by default from \code{survfit} object.
Defines the x-scale end point. In \emph{original} time unit.}

\item{title}{String. Plot title}

\item{title_size}{Numeric. Font size of the title. Defaule = 2.7.}

\item{show_tbls}{Logical. Adds summary tables inside the Kaplan-Meier-plot.}

\item{tbl1_pos}{Numeric Vector length 2 (x, y) for \code{get_surv_times()}.}

\item{tbl2_pos}{Numeric Vector length 2 (x, y) for \code{get_median_table()}.}

\item{followup_times}{Numeric vector of individual length for \code{get_surv_times()}. Defaut = NULL.}

\item{risk_table_size}{Numeric. Font size of the risk table.}

\item{...}{Further arguments forwarded to \code{ggsurvplot()}
(e.g. \code{conf.int = TRUE}, \code{pval = TRUE}, \code{linetype = "strata"}).}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{                   A `ggsurvplot` object list.
}\if{html}{\out{</div>}}
}
\description{
High‑level Kaplan–Meier plot wrapper with add-ons like median & follow-up times table,
table positions, customazation of the scale, title, title-size and more.
}
\examples{
library(survival)
surv <- Surv(time = lung$time, event = lung$status)
fit <- survfit(surv ~ 1, data = lung)
create_surv_plot(lung, fit, xscale = "d_m", x_end = 1300, scale_break = 60.88, time_unit = "Months")
}
