% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_global_tool_path}
\alias{get_global_tool_path}
\title{Returns the path of the folder \code{global_tools}}
\usage{
get_global_tool_path(filename)
}
\arguments{
\item{filename}{String; Access to the file in the \code{global_tools} folder}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{         Path to the file. Example use: `get_global_tool_path("helper_functions.R")`
}\if{html}{\out{</div>}}
}
\description{
Helper Funktion um zu global tools Pfad zu gelangen
}
