% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{get_median_table}
\alias{get_median_table}
\title{(Helper) Median table}
\usage{
get_median_table(fit, time_unit = "Years", xscale = "m_y")
}
\arguments{
\item{fit}{\code{survfit} object.}

\item{time_unit}{String; For the description in the table; e.g. "Years", "Months" or "Days".}

\item{xscale}{String; Transformation specification passed to \code{get_xscale()}.
Character of the form "d_m" (= days to months) or "m_y" (= months to years)
or "m_m" (= remains months).
Allowed pairs (\code{from_to}): d_m, d_y, m_d, m_y, y_d, y_m, d_d, m_m, y_y.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{            A tibble with a single column that is ready to be printed or annotated.
}\if{html}{\out{</div>}}
}
\description{
Median Survival (with 95\% CI) as a one‑row \code{tibble}
}
