% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{get_surv_times}
\alias{get_surv_times}
\title{(Helper) Follow-up times table}
\usage{
get_surv_times(
  fit,
  times = NULL,
  years = c(1, 2, 5),
  time_unit = "Years",
  xscale = "m_y"
)
}
\arguments{
\item{fit}{\code{survfit} object.}

\item{times}{Numeric vector of \emph{original} time scale (e.g. months).}

\item{years}{Default; Numeric vector of time scale in years.}

\item{time_unit}{Label for column header.}

\item{xscale}{Transformation specification passed to \code{get_xscale()}}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{           A table with two columns (follow up survival probabilities (+ confidence intervals)
                    at 1, 2, and 5 years.)
}\if{html}{\out{</div>}}
}
\description{
Survival probabilities at pre‑defined follow‑up times
Follow up Survival probability (in \%)
}
