% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_functions.R
\name{get_xscale}
\alias{get_xscale}
\title{(Helper) Transform the x-achsis of KM curves}
\usage{
get_xscale(xscale = NULL)
}
\arguments{
\item{xscale}{Character of the form "d_m" (= days to months) or "m_y"
(= months to years). See details.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    Numeric factor
}\if{html}{\out{</div>}}
}
\description{
Multiplication factor to transform the x-axis of KM curves
define x-Scale
}
\details{
Allowed pairs (\code{from}_\code{to}): d_m, d_y, m_d, m_y, y_d, y_m.
}
