% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_helper_functions.R
\name{impute_missing_dates}
\alias{impute_missing_dates}
\title{Impute Missing Dates by Replacing "nk" with Standard Day Values}
\usage{
impute_missing_dates(df, raw_date_column, day = "15")
}
\arguments{
\item{df}{Data frame. Data Frame where imputations should take place.}

\item{raw_date_column}{Character. Name of the "raw" date-column, which has the "nk" values. (e.g. \code{"start_date_raw"}).}

\item{day}{Character. Day value, which should be imputed (default is \code{"15"}).}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{            Data frame with imputed day-value column.
}\if{html}{\out{</div>}}
}
\description{
This function imputes date values by replacing \verb{“nk”} with a specified default day (e.g., \code{01}, \code{15}, \code{28}).
The imputed date column is then stored in the DataFrame.
}
