% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_helper_functions.R
\name{parse_mapping}
\alias{parse_mapping}
\title{Parse a RedCap mapping string to a named vector}
\usage{
parse_mapping(mapping_str)
}
\arguments{
\item{mapping_str}{Character string. RedCap mapping string (e.g. "1, Yes | 2, No").}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{         Named character vector: values as names, labels as values.
}\if{html}{\out{</div>}}
}
\description{
Parse a RedCap mapping string to a named vector
}
