% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_helper_functions.R
\name{paste_without_na}
\alias{paste_without_na}
\title{Paste elements together, omitting NAs}
\usage{
paste_without_na(x)
}
\arguments{
\item{x}{Vector. Elements to combine.}
}
\value{
Single character string, elements separated by ", ".
}
\description{
Paste elements together, omitting NAs
combine without NA
}
