% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{read_latest_cleaned_data}
\alias{read_latest_cleaned_data}
\title{Import latest cleaned data}
\usage{
read_latest_cleaned_data(path = file.path("data", "cleaned"), detailed = FALSE)
}
\arguments{
\item{path}{Default path: \code{data/cleaned}}

\item{detailed}{If TRUE, detailed infos are displayed}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     List with all data sets; Access via: `cleaned_data_list[["__NAME__"]]`
                                                or `cleaned_data_list$__NAME__`.
                                                Example use: `result_list <- read_latest_cleaned_data()`
}\if{html}{\out{</div>}}
}
\description{
\code{read_latest_cleaned_data()} returns a list of all cleaned data sets from the directory
\verb{data/cleaned/} making it easier to import them into the central pipeline or report file.
}
