% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_helper_functions.R
\name{reformat_dates}
\alias{reformat_dates}
\title{Reform month & year columns, considering partial dates}
\usage{
reformat_dates(df, base_names)
}
\arguments{
\item{df}{Data frame containing the date columns.}

\item{base_names}{Character vector. Base names (without _month/_year).}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        Data frame with reformatted date columns.
}\if{html}{\out{</div>}}
}
\description{
Reformat dates using month and year columns, or format full dates
for reformatting, considering partial dates
}
