% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{save_cleaned_result}
\alias{save_cleaned_result}
\title{Save recent cleaned data}
\usage{
save_cleaned_result(
  data,
  filename_prefix = "cleaned_data",
  cleaned_dir = "data/cleaned",
  write_csv2 = FALSE,
  create_latest_copy = TRUE
)
}
\arguments{
\item{data}{Data object to be transformed and stored.}

\item{filename_prefix}{String; represents the name of the direcotry to be saved.}

\item{cleaned_dir}{Default: \verb{data/cleaned/} path}

\item{write_csv2}{If = TRUE: CSV with seperator ";"}

\item{create_latest_copy}{If = TRUE: Creates a copy with the name: \code{latest_NAME.rds}}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{               Creates 3 data formats (xlsx, csv, rds) + `latest_NAME.rds`
}\if{html}{\out{</div>}}
}
\description{
\code{save_cleaned_result()} saves results in 3 different data formats (csv, xlsx, rds)
in the directory \verb{data/cleaned/}.
}
\examples{
                  `save_cleaned_result(data, filename_prefix = "therapy_lines")`

}
