% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_helper_functions.R
\name{sort_unique_values}
\alias{sort_unique_values}
\title{Sort unique columns}
\usage{
sort_unique_values(x)
}
\arguments{
\item{x}{Character vector with comma-separated values.}
}
\value{
Character vector with sorted, unique, comma-separated values.
}
\description{
Sort and return unique values in a comma-separated character column
filter unique therapies in column
}
