% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_helper_functions.R
\name{values_to_labels}
\alias{values_to_labels}
\title{Transform raw values to the labels}
\usage{
values_to_labels(df, var)
}
\arguments{
\item{df}{Data frame. The data you want to modify.}

\item{var}{Unquoted variable name in df to convert.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     Data frame with the mapped variable replaced by its label.
}\if{html}{\out{</div>}}
}
\description{
Map coded variable values to their labels using the RedCap data dictionary
Generalized function to map variable values to their labels
}
