% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_helper_functions.R
\name{var_label}
\alias{var_label}
\title{(Helper) Get or set variable label as an attribute.}
\usage{
var_label(x, value = NULL)
}
\arguments{
\item{x}{Vector. The variable to label.}

\item{value}{Character (optional). If supplied, sets the label; if omitted, returns label.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   Label attribute (if value is NULL), or labelled vector (if value is set).
}\if{html}{\out{</div>}}
}
\description{
Get or set a variable label as an attribute
Define the var_label function for setting and getting labels
}
